repeaters_app/requirements.txt
djangorestframework>=3.15.0
